Add new confirmation dialog generic component and delete confirmation for player and team

+++ src/app/features/team/team-list/team-list.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { tap } from 'rxjs/operators';
 
 import { Team } from '../team';
 import { TeamService } from '../team.service';
+import { ConfimationDialogComponent } from 'src/app/shared/confimation-dialog/confimation-dialog.component';
 
 @Component({
   selector: 'app-team-list',
@@ -15,6 +17,7 @@ export class TeamListComponent implements OnInit {
   displayedColumns: string[] = ['Id', 'Name', 'Players', 'Details', 'Delete'];
 
   constructor(
+    public dialog: MatDialog,
     private teamService: TeamService,
   ) { }
 
@@ -28,10 +31,21 @@ export class TeamListComponent implements OnInit {
     );
   }
 
-  deleteTeam(team: Team): void {
-    this.teamService.deleteTeam(team.Id).pipe(
-      tap(() => this.getTeams()),
-    ).subscribe();
+  openDeleteDialog(team: Team): void {
+    const dialogRef = this.dialog.open(ConfimationDialogComponent, {
+      data: {
+        title: 'Delete Team',
+        message: `Are you sure you want to delete the team ${team.Name}?`,
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.teamService.deleteTeam(team.Id).pipe(
+          tap(() => this.getTeams()),
+        ).subscribe();
+      }
+    });
   }
 
 }
